% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_life_time.R
\name{get_life_time}
\alias{get_life_time}
\title{get_retain_users_n_days}
\usage{
get_life_time(ring_retain_new, prediction_retain_one, ring_retain_old,
  life_time_year = 1)
}
\arguments{
\item{ring_retain_new:}{ring_retain_new}

\item{prediction_retain_one:}{prediction_retain_one}

\item{ring_retain_old:}{ring_retain_old}

\item{life_time_year:}{days}
}
\value{
df
}
\description{
get_retain_users_n_days
}
\examples{
get_life_time(10, 7, 30)

}
