% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ring_retain.R
\name{get_ring_retain}
\alias{get_ring_retain}
\title{get_ring_retain}
\usage{
get_ring_retain(start = 0.6, n = 8, max = 0.999)
}
\arguments{
\item{start:}{start}

\item{n:}{n}

\item{max:}{max}
}
\value{
list(ring_retain_new, ring_retain_old)
}
\description{
get_ring_retain
}
\examples{
get_ring_retain()
}
