% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_max.R
\name{if_max}
\alias{if_max}
\title{get_prediction_daily}
\usage{
if_max(days, n)
}
\arguments{
\item{days:}{days}

\item{n:}{n}
}
\value{
n or 0
}
\description{
get_prediction_daily
}
\examples{
if_max(days = 100, n = 10)
}
