% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_df.R
\name{make_df}
\alias{make_df}
\title{制作计算所需数据集}
\usage{
make_df(file_name = "Data/info.csv", train_cr = 0.7, diff_days = 30,
  diff_base = 1.02)
}
\arguments{
\item{file_name:}{用来拟合的文件，包含真实DAU、DNU、次日留存等。默认 "Data/info.csv"}

\item{train_cr:}{全部真实数据中用来做为训练的比例。默认0.8}

\item{diff_days:}{用于计算拟合差异的天数，按截止日倒推，对越近期的拟合成绩越看重。默认30}

\item{diff_base:}{计算差异时，每增加1天，差异(r)按照diff_base加权(r^diff_base)计为当日差异，最终按全部差异的中位数得到总差异。默认1.02}
}
\value{
提供给get_prediction_daily()的参数列表(df_list)，包含拟合数据集(train_df)，测试数据集(test_df)，基于真实数据拟合的天数(train_days)，用于计算拟合差异的天数(diff_days)，计算差异的加权(diff_base)
}
\description{
从csv文件中得到训练/测试数据集，差异天数，差异加权等。
}
\examples{
make_df()
}
