% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit.R
\name{get_fit}
\alias{get_fit}
\title{get_fit}
\usage{
get_fit(no_cores, file_name = "Data/info.csv", n = 50000,
  train_cr = 0.8, diff_days = 30, start = 0.3, max = 0.99999,
  smooth = FALSE)
}
\arguments{
\item{no_cores:}{no_cores}

\item{file_name:}{file_name}

\item{n:}{n}

\item{train_cr:}{train_cr}

\item{diff_days:}{diff_days}

\item{start:}{start}

\item{max:}{max}

\item{smooth:}{smooth}
}
\description{
get_fit
}
\examples{
get_fit()
}
