% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction_daily.R
\name{get_prediction_daily}
\alias{get_prediction_daily}
\title{get_prediction_daily}
\usage{
get_prediction_daily(diff_days = 90, diff_base = 1.02,
  file_name = "info.csv", ring_retain_new = c(0.848, 0.9138, 0.9525,
  0.9679, 0.9801, 0.9861, 0.99, 0.99), ring_retain_old = 0.98,
  prediction_retain_one = 0.48, life_time_year = 1, csv = FALSE,
  plot = TRUE, message = FALSE, smooth = FALSE)
}
\arguments{
\item{diff_days:}{diff_days}

\item{diff_base:}{diff_base}

\item{file_name:}{file_name}

\item{ring_retain_new:}{ring_retain_new}

\item{ring_retain_old:}{ring_retain_old}

\item{prediction_retain_one:}{prediction_retain_one}

\item{life_time_year:}{life_time_year}

\item{csv:}{csv}

\item{plot:}{plot}

\item{message:}{message}

\item{smooth:}{smooth}
}
\value{
df
}
\description{
get_prediction_daily
}
\examples{
get_prediction_daily(file_name = "info.csv")
}
