% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_retain_users_n_days.R
\name{get_retain_users_n_days}
\alias{get_retain_users_n_days}
\title{get_retain_users_n_days}
\usage{
get_retain_users_n_days(retain_users_detail, total_days, start_n, end_n)
}
\arguments{
\item{retain_users_detail:}{retain_users_detail}

\item{total_days:}{total_days}

\item{start_n:}{start_n}

\item{end_n:}{end_n}
}
\value{
df
}
\description{
get_retain_users_n_days
}
\examples{
# get_retain_users_n_days(retain_users_detail = retain_users_detail, 10, 7, 30)
}
