% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_df.R
\name{make_df}
\alias{make_df}
\title{make_df}
\usage{
make_df(file_name = "Data/info.csv", train_cr = 0.7, diff_days = 30,
  diff_base = 1.02)
}
\arguments{
\item{file_name:}{file_name}

\item{train_cr:}{train_cr}

\item{diff_days:}{diff_days}

\item{diff_base:}{diff_base}
}
\value{
df_list
}
\description{
get_prediction_daily
}
\examples{
make_df()
}
